backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
google-genai==0.3.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.7.4
python-multipart==0.0.9
